Moved auto-select to qttools and fixed print statement for real system matrix
